+++ twilight-http/src/request/application/interaction/get_response.rs
@@ -12,6 +12,8 @@ use twilight_model::{
 
 /// Get the original message, by its token.
 ///
+/// This endpoint is not bound to the application's global rate limit.
+///
 /// # Examples
 ///
 /// ```no_run